Split into two tests
