+++ experiments/react-native-uifabric/src/RNTester/TestComponents/Persona/styles.ts
@@ -2,6 +2,7 @@ import { StyleSheet } from 'react-native';
 
 export const styles = StyleSheet.create({
   root: {
+    marginTop: 16,
     flexDirection: 'row'
   },
   settings: {
@@ -25,3 +26,5 @@ export const styles = StyleSheet.create({
 
 export const satyaPhotoUrl =
   'https://www.microsoft.com/en-us/CMSImages/satya.jpg?version=0881eb71-4942-b627-d602-84c832b8a0b6&amp;CollectionId=1b46ce2d-c90d-421e-94f1-cfb6bc6ef6ec';
+
+export const steveBallerPhotoUrl = 'http://www.rmndigital.com/wp-content/uploads/2012/11/stevems.jpg';
